+++ beir_extensions/mlflow.py
@@ -1,0 +1,55 @@
+import os
+import mlflow
+from mlflow.tracking.request_header.abstract_request_header_provider import \
+    RequestHeaderProvider
+
+
+def get_or_create_experiment(name: str) -> str:
+    """
+    Creates mlflow experiment with specified name of
+    retrieves it if it is already created or deleted.
+    :param name: name of the experiment
+    :return: experiment id
+    """
+    client = mlflow.tracking.MlflowClient()
+    experiment = client.get_experiment_by_name(name=name)
+    if experiment is None:
+        experiment_id = client.create_experiment(name=name)
+    else:
+        if dict(experiment)['lifecycle_stage'] == 'deleted':
+            client.restore_experiment(dict(experiment)['experiment_id'])
+        experiment_id = dict(experiment)['experiment_id']
+    return experiment_id
+
+
+def mlflow_flattening(
+    per_dataset_metrics: dict[str, dict[str, float]]
+) -> dict[str, float]:
+    """
+    For given dictionary of dictionaries, flatten the first level and return a
+    a single level dictionary
+    :param per_dataset_metrics: dictionary of dictionaries containing metrics
+    :return: flattened dictionary
+    """
+    return {
+        dataset + '-' + metric: result for dataset in per_dataset_metrics
+        for metric, result in per_dataset_metrics[dataset].items()
+    }
+
+
+class CustomHeaderProvider(RequestHeaderProvider):
+    """
+    Custom header provider
+    """
+    def in_context(self):
+        """Return true for the header to be enabled"""
+        return True
+
+    def request_headers(self):
+        """Define the returned header"""
+        return {
+            'CF-Access-Client-Id': os.environ.get('cf_access_client_id'),
+            'CF-Access-Client-Secret': os.environ.get(
+                'cf_access_client_secret'
+            )
+        }
