+++ beir_extensions/models/base_model.py
