+++ configs/dense_retrieval_experiment.yaml
@@ -28,6 +28,6 @@ model:
   matryoshka_dim: Null  # Null if you don't want to use matryoshka dimension, otherwise select integer value
   score_function: cos_sim
 mlflow:
-  tracking_uri: https://mlflow-demo.atypon.com/
+  tracking_uri: https://rnd-mlflow.literatumonline.com/
   experiment_name: beir-evaluation
   run_name: bge-base-matryoshka-full-instruction
