+++ configs/dense_retrieval_experiment_onnx.yaml
@@ -30,6 +30,6 @@ model:
   sep: ' ' # Seperator between different fields of the documents in corpus eg title, text, etc.
   cls: False  # Set to True if you want to use CLS token for scoring, otherwise set to False
 mlflow:
-  tracking_uri: https://mlflow-demo.atypon.com/
+  tracking_uri: https://rnd-mlflow.literatumonline.com/
   experiment_name: beir-evaluation
   run_name: bge-base-matryoshka-full-instruction
